dsub: Add periodic status update in output (via --summary flag)
PiperOrigin-RevId: 277580242
